Removed callbacks, moved tile_type and module_type to shared location for trace and profiling plugins, Removed parsePollingInterval() function
