+++ src/runtime_src/core/pcie/linux/plugin/xdp/aie_profile.h
@@ -27,11 +27,6 @@ namespace profile {
 
 } // end namespace profile
 
-namespace ctr {
-  void update_device(void* handle);
-  void end_poll(void* handle);
-} // end namespace ctr
-
 } // end namespace aie
 } // end namespace xdp
