+++ src/runtime_src/xdp/profile/plugin/aie_profile_new/aie_profile_plugin.cpp
@@ -55,8 +55,6 @@ namespace xdp {
     db->registerPlugin(this);
     db->registerInfo(info::aie_profile);
     db->getStaticInfo().setAieApplication();
-    // getPollingInterval(); //moved later
-
   }
 
   AieProfilePlugin::~AieProfilePlugin()
@@ -124,9 +122,6 @@ namespace xdp {
     AIEData.deviceID = deviceID;
     AIEData.metadata = std::make_shared<AieProfileMetadata>(deviceID, handle);
     auto& metadata = AIEData.metadata;
-
-    //Get the polling interval after the metadata has been defined.
-    metadata->parsePollingInterval(); 
 
     #ifdef XRT_X86_BUILD
         AIEData.implementation = std::make_unique<AieProfile_x86Impl>(db, AIEData.metadata);
